- update from HEAD

+++ lustre/tests/sanity.sh
@@ -9,8 +9,8 @@
 set -e
 
 ONLY=${ONLY:-"$*"}
-# bug number for skipped test: 13297 2108 9789 3637 9789 3561 12622 12653 12653 5188 10764 16260 20784
-ALWAYS_EXCEPT="                27u   42a  42b  42c  42d  45   51d   65a   65e   68b   75    119d   215 $SANITY_EXCEPT"
+# bug number for skipped test: 13297 2108 9789 3637 9789 3561 12622 12653 12653 5188 10764 16260
+ALWAYS_EXCEPT="                27u   42a  42b  42c  42d  45   51d   65a   65e   68b   75    119d   $SANITY_EXCEPT"
 # bug number for skipped test: 2108 9789 3637 9789 3561 5188/5749 1443
 #ALWAYS_EXCEPT=${ALWAYS_EXCEPT:-"27m 42a 42b 42c 42d 45 68 76"}
 # UPDATE THE COMMENT ABOVE WITH BUG NUMBERS WHEN CHANGING ALWAYS_EXCEPT!
@@ -1828,6 +1828,8 @@ test_36e() {
 run_test 36e "utime on non-owned file (should return error) ===="
 
 test_36f() {
+	local LANG_SAVE=$LANG
+	local LC_LANG_SAVE=$LC_LANG
 	export LANG=C LC_LANG=C # for date language
 
 	DATESTR="Dec 20  2000"
@@ -1848,6 +1850,8 @@ test_36f() {
 		echo "AFTER : $LS_AFTER" && \
 		echo "WANT  : $DATESTR" && \
 		error "$DIR/$tdir/$tfile timestamps changed" || true
+
+	export LANG=$LANG_SAVE LC_LANG=$LC_LANG_SAVE
 }
 run_test 36f "utime on file racing with OST BRW write =========="
 
@@ -3223,6 +3227,7 @@ test_66() {
 run_test 66 "update inode blocks count on client ==============="
 
 LLOOP=
+LLITELOOPLOAD=
 cleanup_68() {
 	trap 0
 	if [ ! -z "$LLOOP" ]; then
@@ -3234,6 +3239,10 @@ cleanup_68() {
 		rm -f $LLOOP
 		unset LLOOP
 	fi
+	if [ ! -z "$LLITELOOPLOAD" ]; then
+		rmmod llite_lloop
+		unset LLITELOOPLOAD
+	fi 
 	rm -f $DIR/f68*
 }
 
@@ -3249,14 +3258,20 @@ swap_used() {
 test_68a() {
 	[ "$UID" != 0 ] && skip_env "must run as root" && return
 
-	grep -q llite_lloop /proc/modules
-	[ $? -ne 0 ] && skip "can't find module llite_lloop" && return
+	trap cleanup_68 EXIT
+
+	if ! module_loaded llite_lloop; then
+		if load_module llite/llite_lloop; then
+			LLITELOOPLOAD=yes
+		else
+			skip_env "can't find module llite_lloop"
+			return
+		fi
+	fi
 
 	LLOOP=$TMP/lloop.`date +%s`.`date +%N`
 	dd if=/dev/zero of=$DIR/f68a bs=4k count=1024
 	$LCTL blockdev_attach $DIR/f68a $LLOOP || error "attach failed"
-
-	trap cleanup_68 EXIT
 
 	directio rdwr $LLOOP 0 1024 4096 || error "direct write failed"
 	directio rdwr $LLOOP 0 1025 4096 && error "direct write should fail"
@@ -6471,13 +6486,13 @@ test_160() {
     # verify contents
     echo "verifying target fid"
     fidc=$($LFS changelog $MDT0 | grep timestamp | grep "CREAT" | \
-	tail -1 | awk '{print $5}')
+	tail -1 | awk '{print $6}')
     fidf=$($LFS path2fid $DIR/$tdir/pics/zach/timestamp)
     [ "$fidc" == "t=$fidf" ] || \
 	err17935 "fid in changelog $fidc != file fid $fidf"
     echo "verifying parent fid"
     fidc=$($LFS changelog $MDT0 | grep timestamp | grep "CREAT" | \
-	tail -1 | awk '{print $6}')
+	tail -1 | awk '{print $7}')
     fidf=$($LFS path2fid $DIR/$tdir/pics/zach)
     [ "$fidc" == "p=$fidf" ] || \
 	err17935 "pfid in changelog $fidc != dir fid $fidf"
@@ -6916,7 +6931,7 @@ test_215() { # for bug 18102
 	TOTAL_LINES=$(cat $TMP/lnet_peers.out |wc -l)
 	OTHER_LINES=$(($TOTAL_LINES - 1))
 	MATCHED_LINES=$(cat $TMP/lnet_peers.out |tail -$TOTAL_LINES |
-			grep -c "^[0-9.]\+@[a-z0-9]\+ *[0-9]\+ *[a-Z]\+ *[0-9]\+ *[0-9]\+ *-\?[0-9]\+ *[0-9]\+ *-\?[0-9]\+ *[0-9]\+$")
+			grep -c "^[0-9.]\+@[a-z0-9]\+ *[0-9]\+ *[a-zA-Z]\+ *[0-9]\+ *[0-9]\+ *-\?[0-9]\+ *[0-9]\+ *-\?[0-9]\+ *[0-9]\+$")
 	[ "$MATCHED_LINES" = "$OTHER_LINES" ] || (cat $TMP/lnet_peers.out && 
 						  error "/proc/sys/lnet/peers misformatted")
 	diff $TMP/lnet_peers.out $TMP/lnet_peers.sys ||
